+++ examples/vn_trader/run.py
@@ -8,41 +8,46 @@ from vnpy_ctp import CtpGateway
 # from vnpy_ctptest import CtptestGateway
 # from vnpy_mini import MiniGateway
 # from vnpy.gateway.minitest import MinitestGateway
-# from vnpy.gateway.femas import FemasGateway
+# from vnpy_femas import FemasGateway
 # from vnpy_sopt import SoptGateway
 # from vnpy.gateway.sopttest import SopttestGateway
-# from vnpy.gateway.sec import SecGateway
+# from vnpy_sec import SecGateway
 # from vnpy_uft import UftGateway
 # from vnpy_esunny import EsunnyGateway
 # from vnpy.gateway.hsoption import HsoptionGateway
 # from vnpy_xtp import XtpGateway
-# from vnpy.gateway.tora import ToraStockGateway
-# from vnpy.gateway.tora import ToraOptionGateway
-# from vnpy.gateway.comstar import ComstarGateway
-# from vnpy.gateway.ib import IbGateway
-# from vnpy.gateway.tap import TapGateway
-# from vnpy.gateway.da import DaGateway
-# from vnpy.gateway.mt5 import Mt5Gateway
+# from vnpy_tora import ToraStockGateway
+# from vnpy_tora import ToraOptionGateway
+# from vnpy_comstar import ComstarGateway
+# from vnpy_ib import IbGateway
+# from vnpy_tap import TapGateway
+# from vnpy_da import DaGateway
 # from vnpy_rohon import RohonGateway
 # from vnpy_tts import TtsGateway
+# from vnpy_sgit import SgitGateway
+# from vnpy_ksgold import KsgoldGateway
+# from vnpy_ost import OstGateway
+# from vnpy_hft import GtjaGateway
+# from vnpy_nhtd import NhFuturesGateway
+# from vnpy_nhtd import NhStockGateway
+# from vnpy_futu import FutuGateway
 
 # from vnpy_ctastrategy import CtaStrategyApp
 # from vnpy_ctabacktester import CtaBacktesterApp
 # from vnpy_spreadtrading import SpreadTradingApp
-# from vnpy.app.algo_trading import AlgoTradingApp
-# from vnpy.app.option_master import OptionMasterApp
-# from vnpy.app.portfolio_strategy import PortfolioStrategyApp
-# from vnpy.app.script_trader import ScriptTraderApp
-# from vnpy.app.market_radar import MarketRadarApp
+# from vnpy_algotrading import AlgoTradingApp
+# from vnpy_optionmaster import OptionMasterApp
+# from vnpy_portfoliostrategy import PortfolioStrategyApp
+# from vnpy_scripttrader import ScriptTraderApp
 # from vnpy_chartwizard import ChartWizardApp
 # from vnpy.app.rpc_service import RpcServiceApp
-# from vnpy.app.excel_rtd import ExcelRtdApp
+# from vnpy_excelrtd import ExcelRtdApp
 # from vnpy_datamanager import DataManagerApp
 # from vnpy_datarecorder import DataRecorderApp
 # from vnpy_riskmanager import RiskManagerApp
 # from vnpy_webtrader import WebTraderApp
-# from vnpy.app.portfolio_manager import PortfolioManagerApp
-# from vnpy.app.paper_account import PaperAccountApp
+# from vnpy_portfoliomanager import PortfolioManagerApp
+# from vnpy_paperaccount import PaperAccountApp
 
 
 def main():
@@ -72,10 +77,15 @@ def main():
     # main_engine.add_gateway(IbGateway)
     # main_engine.add_gateway(TapGateway)
     # main_engine.add_gateway(DaGateway)
-    # main_engine.add_gateway(Mt5Gateway)
     # main_engine.add_gateway(RohonGateway)
-    # main_engine.add_gateway(XgjGateway)
     # main_engine.add_gateway(TtsGateway)
+    # main_engine.add_gateway(SgitGateway)
+    # main_engine.add_gateway(KsgoldGateway)
+    # main_engine.add_gateway(OstGateway)
+    # main_engine.add_gateway(GtjaGateway)
+    # main_engine.add_gateway(NhFuturesGateway)
+    # main_engine.add_gateway(NhStockGateway)
+    # main_engine.add_gateway(FutuGateway)
     
     # main_engine.add_app(CtaStrategyApp)
     # main_engine.add_app(CtaBacktesterApp)
@@ -84,7 +94,6 @@ def main():
     # main_engine.add_app(OptionMasterApp)
     # main_engine.add_app(PortfolioStrategyApp)
     # main_engine.add_app(ScriptTraderApp)
-    # main_engine.add_app(MarketRadarApp)
     # main_engine.add_app(ChartWizardApp)
     # main_engine.add_app(RpcServiceApp)
     # main_engine.add_app(ExcelRtdApp)