+++ vnpy/api/femas/__init__.py
@@ -1,3 +1,1 @@
-from .vnfemasmd import MdApi
-from .vnfemastd import TdApi
-from .femas_constant import *
+from vnpy_femas.api import *