+++ vnpy/api/websocket/__init__.py
@@ -1,1 +1,0 @@
-from .websocket_client import WebsocketClient